\name{ordiconsensus-package}
\alias{ordiconsensus-package}
\alias{ordiconsensus}
\docType{package}
\title{
Consensus RDA package
}
\description{
Package to perform a consensus RDA, a method that finds the consensus among different RDA (or distance-based RDA) performed using different dissimilarity coefficients (e.g. Hellinger, Chord, Bray-Curtis)
}
\details{
\tabular{ll}{
Package: \tab ordiconsensus\cr
Type: \tab Package\cr
Version: \tab 0.3-2\cr
Date: \tab 2012-11-12\cr
License: \tab Unlimited\cr
}
This package includes different functions to perform a consensus RDA. It follows the work of Blanchet et al. (In press)
}
\author{
F. Guillaume Blanchet

Maintainer: F. Guillaume Blanchet <guillaume.blanchet@helsinki.fi>

}
\keyword{ multivariate }
